requirements.txt
asgiref==3.11.0
colorama==0.4.6
coverage==7.13.0
Django==6.0
django-debug-toolbar==6.1.0
factory_boy==3.3.3
Faker==39.0.0
iniconfig==2.3.0
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
pytest==9.0.2
pytest-cov==7.0.0
pytest-django==4.11.1
python-decouple==3.8
sqlparse==0.5.5
tzdata==2025.3
